update ident feature properties
